+++ probe/appclient/multi_client.go
@@ -2,6 +2,7 @@ package appclient
 
 import (
 	"bytes"
+	"compress/gzip"
 	"errors"
 	"fmt"
 	"net/url"
@@ -42,7 +43,7 @@ type MultiAppClient interface {
 	PipeConnection(appID, pipeID string, pipe xfer.Pipe) error
 	PipeClose(appID, pipeID string) error
 	Stop()
-	ReportPublisher
+	Publish(r report.Report) error
 }
 
 // NewMultiAppClient creates a new MultiAppClient.
@@ -155,6 +156,12 @@ func (c *multiClient) Stop() {
 // Publish implements Publisher by publishing the reader to all of the
 // underlying publishers sequentially. To do that, it needs to drain the
 // reader, and recreate new readers for each publisher. Note that it will
+func serializeReport(r report.Report) (*bytes.Buffer, error) {
+	buf := &bytes.Buffer{}
+	err := r.WriteBinary(buf, gzip.DefaultCompression)
+	return buf, err
+}
+
 // publish to one endpoint for each unique ID. Failed publishes don't count.
 func (c *multiClient) Publish(r report.Report) error {
 	c.mtx.Lock()
